Move page ccomponents into documentbody
