+++ package.json
@@ -73,6 +73,7 @@
     "emotion-theming": "^10.0.27",
     "gatsby": "^2.24.63",
     "gatsby-plugin-emotion": "^4.3.11",
+    "gatsby-plugin-google-tagmanager": "^2.3.15",
     "gatsby-plugin-layout": "^1.3.11",
     "gatsby-plugin-react-helmet": "^3.3.11",
     "html-screen-capture-js": "^1.0.47",
